BREAKING [#22] v0.3.0: Fixes the edge case affecting managed inputs
BREAKING: Before this commit it was possible to update global state
directly inside component function body, because all update
notifications were postponed to the next rendering cycle. It turns out
that correct behavior of text cursor in inputs relies on value updates
becoming available within the same cycle, thus we have to change it. The
new behavior is though similar to the local useState(..) hook, which
cannot be used in such way.