v0.8.1: useAsyncData(): patch of SSR behavior
Before: the global state segment associated with the data was
initialized in the non-loading state during SSR, and if the async
operation timed out, the initial state injected into client-side was
non-loading. With React up to 17 it was fine, however React 18+ seems to
run the useEffect() responsible for the client-side loading before
comparing SSR and hydrated layout... thus the previous behavior resulted
in hydration errors (mismatches of server-generated and initial
client-side render).

After: SSR initialization of the global state segment is done with the
"loading" state, which resolves the issue.